+++ CHANGELOG.md
@@ -15,14 +15,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.0.3-alpha] - 2021-12-01
+
+### Added
+
+- Order payloads.
+
 ## [0.0.2-alpha] - 2021-12-01
 
 ### Changed
 
-- Changed Priduct payload name to PluggToProduct.
+- Changed Product payload name to PluggToProduct.
 
 ## [0.0.1-alpha] - 2021-12-01
 
 ### Added
 
-- Plugg To product payload.
+- Plugg To Product payload.